ggplot2 minimal facet theme

ggplot2 minimal facet theme

Usage

theme_facet(
  design = "minimal",
  family = "",
  size = 12,
  size_title = 12,
  size_subtitle = 10,
  size_caption = 10,
  size_strip = 10
)

Arguments

design

"minimal" or "xlt"

Examples

dplyr::starwars |>
  tidyr::drop_na(eye_color, height, mass) |>
  dplyr::filter(mass < 1200) |>
  dplyr::mutate(height = height/100) |>
  ggplot2::ggplot() +
  ggplot2::facet_wrap(~ eye_color) +
  ggplot2::aes(x = mass, y = height, color = "#333333", fill  = "#3399ff") +
  ggplot2::geom_point(shape = 21, stroke = 0.3, alpha = .7, size = 3) +
  ggplot2::scale_y_continuous(
    labels = scales::label_comma(decimal.mark = ",", big.mark = "."),
    limits = c(0, 3)
  ) +
  ggplot2::scale_x_continuous(limits = c(0, 160), breaks = seq(0, 160, 30)) +
  ggplot2::scale_color_identity() +
  ggplot2::scale_fill_identity() +
  ggviz::theme_facet() +
  ggplot2::labs(
    y = "altura (metros)",
    x = "massa (kg)",
    title = "Altura e massa dos personagens de Star Wars por cor dos olhos",
    subtitle = "Aproximadamente 24% dos personagens tem olhos marrons e 9,2% olhos laranjas.\nO personagem mais alto da franquia é o Yarael Poof, com 2,64 metros.\nJabba, por sua vez, é o mais pesado, com 1.358 quilos.",
    caption = "Como o peso do Jabba destoa dos demais, para melhorar a visualização dos dados, optei por excluí-lo.\nDados relativos aos 7 primeiros filmes da franquia."
  ) +
  ggplot2::coord_cartesian(clip = "off", expand = TRUE)
#> Error in loadNamespace(x): there is no package called ‘ggtext’